This script was developed prior to our 8Sept2022 meeting.
This script is an altered version of the PilotDataPrep code used in the June22 BTJR spotlight surveys and was originally used for Andrew’s internship presentation. Used to import, clean, and visualize the BTJR counts from the survey datasheet data per km for August22 surveys (not using specific GPS locations of BTJRs along routes here).

The Pseudocode section is new and will be filled in over this weekend as I re-go through this script to create a new one with the alterations we discussed in Thursday’s meeting (using 300x300m grid cells of the landscape overlaid with the new extended (incidental) routes and individual BTJR GPS locations.